data-migration-mongo/src/main/java/com/gdn/data/migration/mongo/MongoConfiguration.java
package com.gdn.data.migration.mongo;

import org.springframework.boot.context.properties.ConfigurationProperties;
import org.springframework.stereotype.Component;

/**
 * @author Eko Kurniawan Khannedy
 * @since 19/10/16
 */
@ConfigurationProperties(prefix = "mongo")
public class MongoConfiguration {

  private String username;

  private String password;

  private String host;

  private Integer port;

  private String database;

  public String getUsername() {
    return username;
  }

  public void setUsername(String username) {
    this.username = username;
  }

  public String getPassword() {
    return password;
  }

  public void setPassword(String password) {
    this.password = password;
  }

  public String getHost() {
    return host;
  }

  public void setHost(String host) {
    this.host = host;
  }

  public Integer getPort() {
    return port;
  }

  public void setPort(Integer port) {
    this.port = port;
  }

  public String getDatabase() {
    return database;
  }

  public void setDatabase(String database) {
    this.database = database;
  }
}
